\name{GSESeq-package}
\alias{GSESeq-package}
\alias{GSESeq}
\docType{package}

\title{Gene set enrichment for RNA-seq data}

\description{\Sexpr{packageDescription("GSESeq")$Description}}

\author{
  \Sexpr{packageDescription("GSESeq")$Author}

  Maintainer: \Sexpr{packageDescription("GSESeq")$Maintainer}
}

\examples{
## FIXME
}

\keyword{package}